Move project feedback components from /components/resources to /components/studentPortal
