+++ pages/studentPortal/projectFeedback/[feedbackRequestId].tsx
@@ -1,6 +1,6 @@
 import { useRouter } from "next/router";
 import React from "react";
-import FeedbackResponse from "../../../components/resources/projectFeedback/feedbackResponse";
+import FeedbackResponse from "../../../components/studentPortal/projectFeedback/feedbackResponse";
 
 export default function EditFeedbackRequestPage() {
   const router = useRouter();
